Fix example for Nexus
Use the 'nexus' device profile since it's a Nexus we are dealing with. I
don't see how this every worked without it, at least it doesn't for me
on a Nexus 3048.

A bunch of the XML was simply incorrect, like the closing switchport tag
that was missing a leading /. There was an empty
'<switchport></switchport>' which obviously adds nothing.

The enable/disable vlan on a port functions are modified to add and
remove the VLAN instead of trying to replace the entire list of allowed
VLANs. The disable function was previously not working at all as it
tried to write the VLAN to <notallowed> which simply doesn't exist.