Fix error due to use of private paramiko class attribute
This code relies on the internal mapping class in Paramiko, but the variable might hold a normal Python dictionary due to the previous lines, so it crashes when `self._host_keys.lookup(host_port)` is true. The keys to this virtual dictionary are already "x.key.get_name()", so it can be simplified to iterating over it.