Add RPCError accessor for multiple errors
Closes issue #408

When an RPCError represents multiple errors, it is useful for users of the ncclient library to have access to each error, rather than just being able to get a single combined version of the <error-message> tags, and no access to the other fields.

Verified with unit tests, and used in local acceptance tests that use the library to verify NETCONF error return values.