Fix exception on ssh connection with passwordless key
An exception is thrown when trying to connect to a netconf server
with a non-password-protected ssh key. This is because we are trying to
encode an empty string in UTF-8.
This produces this error log:
Authentication failure: AttributeError("'NoneType' object has no attribute 'encode'")
Only encode the password only if it exists.